Split data collection stage, add primary json parsing
Add "data_preprocess" file, and rename data_preparation to data_collection. This splits the raw data collection and storage away from the data preprocessing, which will be easier to follow. The file save is a good staging point.

"data_preprocess": currently processes a single file into a list of price events. Next step is to process multiple files into a dataframe and deduplicate.