Tune-up to data collection
Converted data_collection.py back to script. I had decided late in the process to wrap everything in functions, and while it makes the operations easy to pull in elsewhere, it doesn't help with the speed or code clarity.

Updated interpreter, should make it easier to convince PyCharm to allow figures when I get to that.